hw17, 'teacher' app
