+++ students/forms.py
@@ -1,7 +1,7 @@
 from django.forms import ModelForm
 from django.core.validators import ValidationError
 
-from students.models import Student, Teacher
+from students.models import Student
 
 
 class PersonBaseForm(ModelForm):
@@ -44,9 +44,3 @@ class PersonBaseForm(ModelForm):
 class StudentCreateForm(PersonBaseForm):
     class Meta(PersonBaseForm.Meta):
         fields = ["first_name", "last_name", "email", "phone_number", "birthdate", 'course', 'avatar', 'resume']
-
-
-class TeacherCreateForm(PersonBaseForm):
-    class Meta:
-        model = Teacher
-        fields = ["first_name", "last_name", "email", "phone_number", "course"]
