+++ teachers/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from .models import Teacher
+
 
 # Register your models here.
+admin.site.register(Teacher)
+
