+++ teachers/templates/show_teachers.html
@@ -9,7 +9,7 @@
 
 
 {% block content %}
-<!--    <h1>Students list</h1>-->
+<!--    <h1>Teachers list</h1>-->
 <div class="dropdown">
   <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown"
           aria-expanded="false">Choose a course</button>
@@ -17,7 +17,7 @@
     {% for course in courses_list %}
     <li>
       <a class="dropdown-item"
-         href="http://127.0.0.1:8000/students/teachers/?course={{ course.id }}">{{ course.name }}</a>
+         href="http://127.0.0.1:8000/teachers/?course={{ course.id }}">{{ course.name }}</a>
     </li>
     {% endfor %}
   </ul>
