+++ teachers/urls.py
@@ -14,13 +14,12 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.urls import path
+from teachers.views import get_teachers, create_teacher
 
-from teachers.views import get_teachers, create_teacher, update_teacher
 
 app_name = 'teachers'
 
 urlpatterns = [
     path('', get_teachers, name='list'),
-    path('create/', create_teacher, name='new'),
-    path('update/<int:pk>/', update_teacher, name='edit')
+    path('create/', create_teacher, name='create')
 ]
